refactor: create helper function for reuse another place
